==> scripts/05_Relabel-factors.R <==
# Relabel factors

# 06.02.23 IM changed lens_status_aphakia to lens_status_absent to align with RAAB7 data and script. (_aphakia had been used in error)

library(tidyverse)
library(here)

# Functions 
# Written to recode variables where there are right and left eyes with same variable levels (copied from a help request on R4DS slack...)

# Functions for variables and levels identical across all versions (including DR vars)

history_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:4) %in% levels(x)))
  forcats::fct_recode(x,
                      "eye_history_not_blind" = "1",
                      "eye_history_blind_cataract" = "2",
                      "eye_history_blind_other" = "3",
                      "eye_history_surgery_cataract" = "4"
  )
}


spectacles_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:2) %in% levels(x)))
  forcats::fct_recode(x,
                      "false" = "1",
                      "true" = "2"
  )
  
}


lens_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:6) %in% levels(x)))
  forcats::fct_recode(x,
                      "lens_status_normal" = "1",
                      "lens_status_opacity" = "2",
                      "lens_status_absent" = "3",
                      "lens_status_pseudophakia_no_pco" = "4",
                      "lens_status_pseudophakia_with_pco" = "5",
                      "lens_status_no_view" = "6"
  )
}

place_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:5) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_place_gov_hospital" = "1",
                      "surgery_place_voluntary_hospital" = "2",
                      "surgery_place_private_hospital" = "3",
                      "surgery_place_camp_improvised" = "4",
                      "surgery_place_traditional" = "5"
  )
}

cost_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:3) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_cost_totally_free" = "1",
                      "surgery_cost_partially_free" = "2",
                      "surgery_cost_fully_paid" = "3"
  )
}

type_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:3) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_type_non_iol" = "1",
                      "surgery_type_iol" = "2",
                      "surgery_type_couching" = "3"
  ) %>% 
    forcats::fct_drop(only = "0")
}

drmethod_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:2) %in% levels(x)))
  forcats::fct_recode(x,
                      "dr_retinopathy_method_dilatation_fundoscopy" = "1",
                      "dr_retinopathy_method_refused" = "2"
  )
}


drret_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:6) %in% levels(x)))
  forcats::fct_recode(x,
                      "dr_retinopathy_grade_none" = "1",
                      "dr_retinopathy_grade_mild" = "2",
                      "dr_retinopathy_grade_observable" = "3",
                      "dr_retinopathy_grade_referable" = "4",
                      "dr_retinopathy_grade_proliferative" = "5",
                      "dr_retinopathy_grade_not_visualised" = "6"
  )
}

drmac_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:4) %in% levels(x)))
  forcats::fct_recode(x,
                      "dr_maculopathy_grade_none" = "1",
                      "dr_maculopathy_grade_observable" = "2",
                      "dr_maculopathy_grade_referable" = "3",
                      "dr_maculopathy_grade_not_visualised" = "4"
  )
}

drscar_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:5) %in% levels(x)))
  forcats::fct_recode(x,
                      "dr_laser_photocoagulation_scars_absent" = "1",
                      "dr_laser_photocoagulation_scars_pan_retinal" = "2",
                      "dr_laser_photocoagulation_scars_macular" = "3",
                      "dr_laser_photocoagulation_scars_pan_retinal_and_macular" = "4",
                      "dr_laser_photocoagulation_scars_not_visualised" = "5"
  )
}

# Functions requiring modified code per RAAB version

# VA levels are integers and need to stay as integers

va_recode6 <- function(x) {
  stopifnot(is.numeric(x))
  recode(x,
         `1` = 0.3,
         `2` = 0.47,
         `3` = 1,
         `4` = 1.3,
         `5` = 1.8,
         `6` = 3,
         `7` = 4)
}

va_recode5 <- function(x) {
  stopifnot(is.numeric(x))
  recode(x,
         `1` = 0.47,
         `2` = 1,
         `3` = 1.3,
         `4` = 1.8,
         `5` = 3,
         `6` = 4)
}


# Need to code for RACSS data sets without a level 8 for oncho
cause_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:14) %in% levels(x)))
  forcats::fct_recode(x,
                      "poor_vision_cause_uncorrected_refractive_error" = "1",
                      "poor_vision_cause_aphakia_uncorrected" = "2",
                      "poor_vision_cause_cataract_untreated" = "3",
                      "poor_vision_cause_cataract_surgical_complications" = "4",
                      "poor_vision_cause_trachomatous_corneal_opacity" = "5",
                      "poor_vision_cause_other_corneal_opacity" = "6",
                      "poor_vision_cause_phthisis" = "7",
                      "poor_vision_cause_onchocerciasis" = "8",
                      "poor_vision_cause_glaucoma" = "9",
                      "poor_vision_cause_diabetic_retinopathy" = "10",
                      "poor_vision_cause_age_related_macular_degeneration" = "11",
                      "poor_vision_cause_other_posterior_segment_disease" = "12",
                      "poor_vision_cause_other_globe_or_cns_abnormalities" = "13",
                      "poor_vision_cause_not_examined" = "14"
  )
}

# Need to confirm that level 7 (other) is the missing level in RACSS
barrier_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:7) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_none_reason_unnecessary" = "1",
                      "surgery_none_reason_fear" = "2",
                      "surgery_none_reason_cost" = "3",
                      "surgery_none_reason_denied" = "4",
                      "surgery_none_reason_unaware" = "5",
                      "surgery_none_reason_access" = "6",
                      "surgery_none_reason_other" = "7",
  )
}

# Check how the post op reasons from RACSS surveys were translated to RAAB5 levels
postopreason_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:5) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_poor_vision_reason_ocular_comorbidity" = "1",
                      "surgery_poor_vision_reason_operative_complications" = "2",
                      "surgery_poor_vision_reason_uncorrected_refractive_error" = "3",
                      "surgery_poor_vision_reason_longterm_complications" = "4",
                      "surgery_poor_vision_reason_not_applicable" = "5"
  )
}


# RAAB6

# Variables were not factors from previous steps, converting to factors and dropping the level = 0.  
raab6_sur_data_newvarslevs <- read.csv(here('outputs', 'raab6_sur_data_newvars.csv')) %>% 
  
  mutate(across(c('gbd_superreg', 'gender', 'exam_status',
                  'eye_history_right', 'eye_history_left',
                  'spectacles_used_distance', 'spectacles_used_near', 
                  'lens_status_right', 'lens_status_left',
                  'poor_vision_cause_right', 'poor_vision_cause_left', 'poor_vision_cause_principle',
                  'surgery_none_reason_1', 'surgery_none_reason_2',
                  'surgery_place_right', 'surgery_place_left',
                  'surgery_type_right', 'surgery_type_left',
                  'surgery_cost_right', 'surgery_cost_left',
                  'surgery_poor_vision_reason_right', 'surgery_poor_vision_reason_left',
                  'dr_diabetes_known', 'dr_diabetes_blood_consent',
                  'dr_diabetic_treatment', 'dr_diabetic_last_exam', 
                  'dr_retinopathy_method_right', 'dr_retinopathy_method_left',
                  'dr_retinopathy_grade_right', 'dr_maculopathy_grade_right', 'dr_laser_photocoagulation_scars_right',
                  'dr_retinopathy_grade_left', 'dr_maculopathy_grade_left', 'dr_laser_photocoagulation_scars_left'),
                as.factor)) %>% 
  
  # Data sets downloaded from Repo used 0 where there was no observation required - replacing these with NA  
  # For DR vars: 'knowndm' used 0 for missing/ not examined along with 1 or 2, 'noblood' and 'drnodilate' were logical with blank cells instead of 0 values
  mutate(across(where(is.factor), ~replace(., . == "0", NA))) %>% 

  # Drop empty level 0 from all factors
  droplevels() %>% 
  
  # Recoding factor levels to change numeric codes to RAAB7 strings using functions above or code below
mutate(gbd_superreg = fct_recode(gbd_superreg, 
                                 "sea_ea_o" = "gbd1",
                                 "sa" = "gbd2",
                                 "ce_ee_ca" = "gbd3",
                                 "na_me" = "gbd4",
                                 "ssa" = "gbd5",
                                 "lac" = "gbd6",
                                 "hi" = "gbd7"
)) %>%
  
  mutate(exam_status = fct_recode(exam_status, 
                                  "exam_status_examined" = "1",
                                  "exam_status_unavailable" = "2",
                                  "exam_status_refused" = "3",
                                  "exam_status_incapable" = "4"
  )) %>%
  
  mutate(gender = fct_recode(gender,
                             "male" = "1",
                             "female" = "2")) %>%
  
  mutate(
    across(c(eye_history_right, eye_history_left),
           history_recode
    )) %>% 
  
  mutate(
    across(c(spectacles_used_distance, spectacles_used_near),
           spectacles_recode
    )) %>% 
  
  mutate(
    across(c(right_distance_acuity_presenting, left_distance_acuity_presenting,right_distance_acuity_pinhole, left_distance_acuity_pinhole),
           va_recode6
    )) %>%
  
  mutate(
    across(c(lens_status_right, lens_status_left),
           lens_recode
    )) %>% 
  
  mutate(
    across(c(poor_vision_cause_right, poor_vision_cause_left, poor_vision_cause_principle),
           cause_recode
    )) %>%
  
  mutate(
    across(c(surgery_none_reason_1, surgery_none_reason_2),
           barrier_recode
    )) %>% 
  
  mutate(
    across(c(surgery_place_right, surgery_place_left),
           place_recode
    )) %>% 
  
  mutate(
    across(c(surgery_cost_right, surgery_cost_left),
           cost_recode
    )) %>% 
  
  mutate(
    across(c(surgery_type_right, surgery_type_left),
           type_recode
    )) %>% 
  
  mutate(
    across(c(surgery_poor_vision_reason_right, surgery_poor_vision_reason_left),
           postopreason_recode
    )) %>% 
  
  mutate(dr_diabetes_known = fct_recode(dr_diabetes_known,
                                            "false" = "1",
                                            "true" = "2"
  )) %>% 
  
  mutate(dr_diabetes_blood_consent = fct_recode(dr_diabetes_blood_consent, #old RAAB var framed as a negative "noblood" so swapping t and f
                                            "true" = "FALSE",
                                            "false"  = "TRUE"
  )) %>%
  
  mutate(dr_diabetic_treatment = fct_recode(dr_diabetic_treatment,
                                         "dr_diabetic_treatment_none" = "1",
                                         "dr_diabetic_treatment_diet" = "2",
                                         "dr_diabetic_treatment_tablets" = "3",
                                         "dr_diabetic_treatment_insulin" = "4",
                                         "dr_diabetic_treatment_tablets_insulin" = "5",
                                         "dr_diabetic_treatment_other" = "6"
  )) %>%
  
  mutate(dr_diabetic_last_exam = fct_recode(dr_diabetic_last_exam,
                                            "dr_diabetic_last_exam_none" = "1",
                                            "dr_diabetic_last_exam_0_12_months" = "2",
                                            "dr_diabetic_last_exam_13_24_months" = "3",
                                            "dr_diabetic_last_exam_over_24_months" = "4"
  )) %>%
  
  mutate(
    across(c(dr_retinopathy_method_right, dr_retinopathy_method_left),
           drmethod_recode
    )) %>%
  
  mutate(
    across(c(dr_retinopathy_grade_right, dr_retinopathy_grade_left),
           drret_recode
    )) %>% 
  
  mutate(
    across(c(dr_maculopathy_grade_right, dr_maculopathy_grade_left),
           drmac_recode
    )) %>%
  
  mutate(
    across(c(dr_laser_photocoagulation_scars_right, dr_laser_photocoagulation_scars_left),
           drscar_recode
    )) 


write.csv(raab6_sur_data_newvarslevs, here('outputs', 'raab6_sur_data_newvarslevs.csv'), row.names = FALSE)


# RAAB5

# The only difference between RAAB5 and RAAB6 datasets is the change in VA measurement

raab5_sur_data_newvarslevs <- read.csv(here('outputs', 'raab5_sur_data_newvars.csv')) %>% 
  
  mutate(across(c('gbd_superreg', 'gender', 'exam_status',
                  'eye_history_right', 'eye_history_left',
                  'spectacles_used_distance', 'spectacles_used_near', 
                  'lens_status_right', 'lens_status_left',
                  'poor_vision_cause_right', 'poor_vision_cause_left', 'poor_vision_cause_principle',
                  'surgery_none_reason_1', 'surgery_none_reason_2',
                  'surgery_place_right', 'surgery_place_left',
                  'surgery_type_right', 'surgery_type_left',
                  'surgery_cost_right', 'surgery_cost_left',
                  'surgery_poor_vision_reason_right', 'surgery_poor_vision_reason_left',
                  'dr_diabetes_known', 'dr_diabetes_blood_consent',
                  'dr_diabetic_treatment', 'dr_diabetic_last_exam',
                  'dr_retinopathy_method_right', 'dr_retinopathy_method_left',
                  'dr_retinopathy_grade_right', 'dr_maculopathy_grade_right', 'dr_laser_photocoagulation_scars_right',
                  'dr_retinopathy_grade_left', 'dr_maculopathy_grade_left', 'dr_laser_photocoagulation_scars_left'),
                as.factor)) %>% 
  
  # Data sets downloaded from Repo used 0 where there was no observation required - replacing these with NA  
  mutate(across(where(is.factor), ~replace(., . == "0", NA))) %>% 
  # Drop empty level 0 from all factors
  droplevels() %>% 
  
  # Recoding factor levels to change numeric codes to RAAB7 strings using functions above or code below
  mutate(gbd_superreg = fct_recode(gbd_superreg, 
                                   "sea_ea_o" = "gbd1",
                                   "sa" = "gbd2",
                                   "ce_ee_ca" = "gbd3",
                                   "na_me" = "gbd4",
                                   "ssa" = "gbd5",
                                   "lac" = "gbd6",
                                   "hi" = "gbd7"
  )) %>%
  
  mutate(exam_status = fct_recode(exam_status, 
                                  "exam_status_examined" = "1",
                                  "exam_status_unavailable" = "2",
                                  "exam_status_refused" = "3",
                                  "exam_status_incapable" = "4"
  )) %>%
  
  mutate(gender = fct_recode(gender,
                             "male" = "1",
                             "female" = "2")) %>%
  
  mutate(
    across(c(eye_history_right, eye_history_left),
           history_recode
    )) %>% 
  
  mutate(
    across(c(spectacles_used_distance, spectacles_used_near),
           spectacles_recode
    )) %>% 
  
  mutate(
    across(c(right_distance_acuity_presenting, left_distance_acuity_presenting,right_distance_acuity_pinhole, left_distance_acuity_pinhole),
           va_recode5
    )) %>%
  
  mutate(
    across(c(lens_status_right, lens_status_left),
           lens_recode
    )) %>% 
  
  mutate(
    across(c(poor_vision_cause_right, poor_vision_cause_left, poor_vision_cause_principle),
           cause_recode
    )) %>%
  
  mutate(
    across(c(surgery_none_reason_1, surgery_none_reason_2),
           barrier_recode
    )) %>% 
  
  mutate(
    across(c(surgery_place_right, surgery_place_left),
           place_recode
    )) %>% 
  
  mutate(
    across(c(surgery_cost_right, surgery_cost_left),
           cost_recode
    )) %>% 
  
  mutate(
    across(c(surgery_type_right, surgery_type_left),
           type_recode
    )) %>% 
  
  mutate(
    across(c(surgery_poor_vision_reason_right, surgery_poor_vision_reason_left),
           postopreason_recode
    )) %>% 
  
  mutate(dr_diabetes_known = fct_recode(dr_diabetes_known,
                                        "false" = "1",
                                        "true" = "2"
  )) %>% 
  
  mutate(dr_diabetes_blood_consent = fct_recode(dr_diabetes_blood_consent, #old RAAB var framed as a negative "noblood" so swapping t and f
                                                "true" = "FALSE",
                                                "false"  = "TRUE"
  )) %>%
  
  mutate(dr_diabetic_treatment = fct_recode(dr_diabetic_treatment,
                                            "dr_diabetic_treatment_none" = "1",
                                            "dr_diabetic_treatment_diet" = "2",
                                            "dr_diabetic_treatment_tablets" = "3",
                                            "dr_diabetic_treatment_insulin" = "4",
                                            "dr_diabetic_treatment_tablets_insulin" = "5",
                                            "dr_diabetic_treatment_other" = "6"
  )) %>%
  
  mutate(dr_diabetic_last_exam = fct_recode(dr_diabetic_last_exam,
                                            "dr_diabetic_last_exam_none" = "1",
                                            "dr_diabetic_last_exam_0_12_months" = "2",
                                            "dr_diabetic_last_exam_13_24_months" = "3",
                                            "dr_diabetic_last_exam_over_24_months" = "4"
  )) %>%
  
  mutate(
    across(c(dr_retinopathy_method_right, dr_retinopathy_method_left),
           drmethod_recode
    )) %>%
  
  mutate(
    across(c(dr_retinopathy_grade_right, dr_retinopathy_grade_left),
           drret_recode
    )) %>% 
  
  mutate(
    across(c(dr_maculopathy_grade_right, dr_maculopathy_grade_left),
           drmac_recode
    )) %>%
  
  mutate(
    across(c(dr_laser_photocoagulation_scars_right, dr_laser_photocoagulation_scars_left),
           drscar_recode
    )) 

write.csv(raab5_sur_data_newvarslevs, here('outputs', 'raab5_sur_data_newvarslevs.csv'), row.names = FALSE)


# fct_count(raab5_sur_data_newvarslevs$Poor.Vision.Cause.Right)

# RACSS 

# RACSS are all stored with the same variable cols as RAAB5 but factor levels differ
# VA to be relabeled the same way as RAAB5
# exam status only coded 1-3, not 1-4
# barriers to surgery only coded 1-6, not 1-7; need HL to explain difference, assume 7 is dropped from conversion
# oncho not present in RACSS so no cause 8 in data set - code stopifnot breaks in the absence of a level 8?
# confirm with HL how the postop reasons for poor vision determined
# not using any of the DR recodes as no DR data in the RACSS dataset


# va_recode_racss <- function(x) {
#   stopifnot(is.factor(x))
#   stopifnot(all((1:6) %in% levels(x)))
#   forcats::fct_recode(x,
#                       "can_see_at_least_618" = "1",
#                       "cant_see_618_can_see_660" = "2",
#                       "cant_see_618_can_see_360" = "3",
#                       "cant_see_360_can_see_160" = "4",
#                       "light_perception" = "5",
#                       "no_perception_light" = "6"
#   )
# }


cause_recode_racss <- function(x) {
  stopifnot(is.factor(x))
  all((1:14) %in% levels(x))
  forcats::fct_recode(x,
                      "poor_vision_cause_uncorrected_refractive_error" = "1",
                      "poor_vision_cause_aphakia_uncorrected" = "2",
                      "poor_vision_cause_cataract_untreated" = "3",
                      "poor_vision_cause_cataract_surgical_complications" = "4",
                      "poor_vision_cause_trachomatous_corneal_opacity" = "5",
                      "poor_vision_cause_other_corneal_opacity" = "6",
                      "poor_vision_cause_phthisis" = "7",
                      "poor_vision_cause_glaucoma" = "9",
                      "poor_vision_cause_diabetic_retinopathy" = "10",
                      "poor_vision_cause_age_related_macular_degeneration" = "11",
                      "poor_vision_cause_other_posterior_segment_disease" = "12",
                      "poor_vision_cause_other_globe_or_cns_abnormalities" = "13",
                      "poor_vision_cause_not_examined" = "14"
  )
}


barrier_recode_racss <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:6) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_none_reason_unnecessary" = "1",
                      "surgery_none_reason_fear" = "2",
                      "surgery_none_reason_cost" = "3",
                      "surgery_none_reason_denied" = "4",
                      "surgery_none_reason_unaware" = "5",
                      "surgery_none_reason_access" = "6"
  )
}

# barrier_recode needs updated for RACSS as coding is 1-6, not 1-7

postopreason_recode <- function(x) {
  stopifnot(is.factor(x))
  stopifnot(all((1:5) %in% levels(x)))
  forcats::fct_recode(x,
                      "surgery_poor_vision_reason_ocular_comorbidity" = "1",
                      "surgery_poor_vision_reason_operative_complications" = "2",
                      "surgery_poor_vision_reason_uncorrected_refractive_error" = "3",
                      "surgery_poor_vision_reason_longterm_complications" = "4",
                      "surgery_poor_vision_reason_not_applicable" = "5"
  )
}


racss_sur_data_newvarslevs <- read.csv(here('outputs', 'racss_sur_data_newvars.csv')) %>% 
  
  mutate(across(c('gbd_superreg', 'gender', 'exam_status',
                  'eye_history_right', 'eye_history_left',
                  'spectacles_used_distance', 'spectacles_used_near', 
                  'lens_status_right', 'lens_status_left',
                  'poor_vision_cause_right', 'poor_vision_cause_left', 'poor_vision_cause_principle',
                  'surgery_none_reason_1', 'surgery_none_reason_2',
                  'surgery_place_right', 'surgery_place_left',
                  'surgery_type_right', 'surgery_type_left',
                  'surgery_cost_right', 'surgery_cost_left',
                  'surgery_poor_vision_reason_right', 'surgery_poor_vision_reason_left',
                  'dr_diabetes_known', 'dr_diabetes_blood_consent',
                  'dr_diabetic_treatment', 'dr_diabetic_last_exam',
                  'dr_retinopathy_method_right', 'dr_retinopathy_method_left',
                  'dr_retinopathy_grade_right', 'dr_maculopathy_grade_right', 'dr_laser_photocoagulation_scars_right',
                  'dr_retinopathy_grade_left', 'dr_maculopathy_grade_left', 'dr_laser_photocoagulation_scars_left'),
                as.factor)) %>% 
  
  # Data sets downloaded from Repo used 0 where there was no observation required - replacing these with NA  
  mutate(across(where(is.factor), ~replace(., . == "0", NA))) %>% 
  # Drop empty level 0 from all factors
  droplevels() %>% 
  
  # Recoding factor levels to change numeric codes to RAAB7 strings
  mutate(gbd_superreg = fct_recode(gbd_superreg, 
                                   "sea_ea_o" = "gbd1",
                                   "sa" = "gbd2",
                                   "ce_ee_ca" = "gbd3",
                                   "na_me" = "gbd4",
                                   "ssa" = "gbd5",
                                   "lac" = "gbd6",
                                   "hi" = "gbd7"
  )) %>%
  
  mutate(exam_status = fct_recode(exam_status, 
                                  "exam_status_examined" = "1",
                                  "exam_status_unavailable" = "2",
                                  "exam_status_refused" = "3"
  )) %>%
  
  mutate(gender = fct_recode(gender,
                             "male" = "1",
                             "female" = "2")) %>%
  
  mutate(
    across(c(eye_history_right, eye_history_left),
           history_recode
    )) %>% 
  
  mutate(spectacles_used_distance = fct_recode(spectacles_used_distance,
                                               "false" = "1",
                                               "true" = "2")) %>%
  
  mutate(
    across(c(right_distance_acuity_presenting, left_distance_acuity_presenting,right_distance_acuity_pinhole, left_distance_acuity_pinhole),
           va_recode5
    )) %>%
  
  mutate(
    across(c(lens_status_right, lens_status_left),
           lens_recode
    )) %>% 
  
  mutate(
    across(c(poor_vision_cause_right, poor_vision_cause_left, poor_vision_cause_principle),
           cause_recode_racss
    )) %>%
  
  mutate(
    across(c(surgery_none_reason_1, surgery_none_reason_2),
           barrier_recode_racss
    )) %>% 
  
  mutate(
    across(c(surgery_place_right, surgery_place_left),
           place_recode
    )) %>% 
  
  mutate(
    across(c(surgery_cost_right, surgery_cost_left),
           cost_recode
    )) %>% 
  
  mutate(
    across(c(surgery_type_right, surgery_type_left),
           type_recode
    )) %>% 
  
  mutate(
    across(c(surgery_poor_vision_reason_right, surgery_poor_vision_reason_left),
           postopreason_recode
    ))

write.csv(racss_sur_data_newvarslevs, here('outputs', 'racss_sur_data_newvarslevs.csv'), row.names = FALSE)
